+++ lib/view_models/privacy_and_security_settings_model.dart
@@ -1,32 +1,61 @@
+import 'package:cccc/services/local_authentication_service.dart';
+import 'package:cccc/services/logger_init.dart';
+import 'package:cccc/services/shared_preference_service.dart';
 import 'package:flutter/material.dart';
-import 'package:local_auth/local_auth.dart';
-import 'package:cccc/extensions/string_extension.dart';
 
 class PrivacyAndSecuritySettingsModel with ChangeNotifier {
-  final LocalAuthentication _auth = LocalAuthentication();
+  PrivacyAndSecuritySettingsModel({required this.localAuth});
+
+  final LocalAuthenticationService localAuth;
+  final SharedPreferencesService _prefs = SharedPreferencesService();
 
   /// String value of the first element of the `availableBioMetrics` list
   String get availableBiometric => _availableBiometric;
 
-  /// boolean value on whether the device is able to check Biometrics.
-  /// Defaults to `false`
-  bool get canCheckBiometrics => _canCheckBiometrics;
+  /// [bool] value for [Switch] widget in `LocalAuthenticationListTile` class
+  bool get switchValue => _switchValue;
 
   late String _availableBiometric;
-  late bool _canCheckBiometrics;
+  bool _switchValue = false;
 
-  /// A function that checks if the user's device supports Biometric
-  /// Authentication. It updates the value `_canCheckBiometrics` and
-  /// `_availableBiometric` based on the result.
-  void checkBiometric() async {
-    final canCheckBiometrics = await _auth.canCheckBiometrics;
+  /// A function that gets the [String] value of availableBiometric from
+  /// [LocalAuthenticationService] class.
+  Future<String> getAvailableBiometrics() async {
+    logger.d('getAvailableBiometrics future called');
 
-    if (canCheckBiometrics) {
-      final availableBioMetrics = await _auth.getAvailableBiometrics();
+    return await localAuth.getAvailableBiometric();
+  }
 
-      _canCheckBiometrics = canCheckBiometrics;
-      _availableBiometric = availableBioMetrics.first.name.title;
-      notifyListeners();
+  /// A function that gets [bool] value for [_switchValue] from [LocalAuthenticationService]
+  /// class.
+  Future<void> getSwitchValue() async {
+    final switchValue = await localAuth.getUseLocalAuth();
+    _switchValue = switchValue;
+    logger.d('_switchValue after future $_switchValue');
+  }
+
+  Future<void> switchOnChange(BuildContext context, bool value) async {
+    logger.d('switch value: $value');
+
+    _switchValue = value;
+    notifyListeners();
+
+    if (value) {
+      final authResult = await localAuth.authenticate();
+      final authSuccessful = authResult['authenticated'] as bool;
+
+      if (authSuccessful) {
+        await _prefs.update('useLocalAuth', value);
+      } else {
+        _switchValue = false;
+        notifyListeners();
+      }
+    } else {
+      final updated = await _prefs.update('useLocalAuth', value);
+
+      if (updated) {
+        localAuth.setIsAuthenticated(false);
+      }
     }
   }
 }