+++ lib/widgets/local_authentication_list_tile.dart
@@ -1,53 +1,35 @@
 import 'package:cccc/providers.dart'
-    show
-        localAuthenticationServiceProvider,
-        privacyAndSecuritySettingsModelProvider;
+    show privacyAndSecuritySettingsModelProvider;
 import 'package:cccc/services/logger_init.dart';
-import 'package:cccc/styles/text_styles.dart';
+import 'package:cccc/widgets/custom_future_builder.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-class LocalAuthenticationListTile extends ConsumerStatefulWidget {
+import 'local_authentication_switch.dart';
+
+class LocalAuthenticationListTile extends StatelessWidget {
   const LocalAuthenticationListTile({Key? key})
       : super(key: const ValueKey('LocalAuthenticationListTile'));
 
   @override
-  ConsumerState<ConsumerStatefulWidget> createState() =>
-      _LocalAuthenticationListTileState();
-}
-
-class _LocalAuthenticationListTileState
-    extends ConsumerState<LocalAuthenticationListTile> {
-  @override
-  void initState() {
-    super.initState();
-    logger.d('inside LocalAuthenticationListTile setState');
-
-    ref.read(privacyAndSecuritySettingsModelProvider).checkBiometric();
-  }
-
-  @override
   Widget build(BuildContext context) {
-    final viewModel = ref.watch(privacyAndSecuritySettingsModelProvider);
-    final localAuth = ref.watch(localAuthenticationServiceProvider);
+    logger.d('[LocalAuthenticationListTile] widget building...');
 
-    if (viewModel.canCheckBiometrics) {
-      return ListTile(
-        title: const Text('Biometric Authentication'),
-        subtitle: Text(
-          viewModel.availableBiometric,
-          style: TextStyles.captionGreyBold,
-        ),
-        trailing: Switch.adaptive(
-          value: localAuth.useLocalAuth,
-          onChanged: (value) => localAuth.useAuthenticationOnChaged(
-            context,
-            value,
-          ),
-        ),
-      );
-    } else {
-      return const SizedBox.shrink();
-    }
+    return ListTile(
+      title: const Text('Biometric Authentication'),
+      subtitle: Consumer(
+        builder: (context, ref, child) {
+          return CustomFutureBuilder<String>(
+            future: ref
+                .read(privacyAndSecuritySettingsModelProvider)
+                .getAvailableBiometrics(),
+            builder: (context, availableBiometric) => Text(availableBiometric!),
+            errorBuilder: (context, e) => Text(e.toString(), maxLines: 1),
+            loadingWidget: const Text('Loading...'),
+          );
+        },
+      ),
+      trailing: const LocalAuthenticationSwitch(),
+    );
   }
 }